fix up scheduling bugs, test static and spectrum

+++ src/color_cycling.hh
@@ -14,5 +14,5 @@ public:
 private:
     uint8_t decColor;
     uint8_t incColor;
-    uint8_t iColorIndex;
+    int16_t iColorIndex;
 };
